refactor spill benchmark
